[node-w3capi] extract api and rendering into two separate modules
