+++ node-w3capi/package.json
@@ -2,6 +2,7 @@
   "name": "w3c",
   "version": "1.0.0",
   "description": "",
+  "type": "module",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
